rename hierarchy folder
