+++ core/src/main/java/tachyon/worker/tiered/StorageDir.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon.worker.hierarchy;
+package tachyon.worker.tiered;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
