+++ core/src/test/java/tachyon/worker/tiered/EvictStrategyTest.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package tachyon.worker.hierarchy;
+package tachyon.worker.tiered;
 
 import java.io.File;
 import java.io.IOException;
@@ -83,7 +83,7 @@ public class EvictStrategyTest {
     EvictStrategy eviction = new EvictLRU(true);
 
     pinList.add(1);
-    Pair<StorageDir, List<tachyon.worker.hierarchy.BlockInfo>> lruResult =
+    Pair<StorageDir, List<tachyon.worker.tiered.BlockInfo>> lruResult =
         eviction.getDirCandidate(mStorageDirs, pinList, 300);
     Assert.assertEquals(mStorageDirs[1], lruResult.getFirst());
     Assert.assertEquals(1, lruResult.getSecond().size());
@@ -118,7 +118,7 @@ public class EvictStrategyTest {
     EvictStrategy eviction = new EvictPartialLRU(true);
 
     pinList.add(1);
-    Pair<StorageDir, List<tachyon.worker.hierarchy.BlockInfo>> lruResult =
+    Pair<StorageDir, List<tachyon.worker.tiered.BlockInfo>> lruResult =
         eviction.getDirCandidate(mStorageDirs, pinList, 600);
     Assert.assertEquals(mStorageDirs[1], lruResult.getFirst());
     Assert.assertEquals(1, lruResult.getSecond().size());
