+++ core/src/test/java/tachyon/worker/tiered/HierarchyStoreTest.java
@@ -13,7 +13,7 @@
  * the License.
  */
 
-package tachyon.worker.hierarchy;
+package tachyon.worker.tiered;
 
 import java.io.IOException;
