+++ core/src/test/java/tachyon/worker/tiered/StorageTierTest.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package tachyon.worker.hierarchy;
+package tachyon.worker.tiered;
 
 import java.io.File;
 import java.io.IOException;
